fix(eval): Separate streamlit-dependent utils
